A big code scrub, applying a uniform coding style and updating copyrights
Applies the GNU coding style across the project and ensures each source
file has a copyright header.

Also this rips out the complex code for spawing processes (we now just
leave it up to gdb instead). This is potentially less flexible, and can be
revisited later, but for now it's much simpler and works!